refactor: pluralize the mesh directories to meshes/
The mesh headers lived under the singular `numerics/mesh/` and a loose `entities/mesh_entity.h`, inconsistent with the pluralized category directories used elsewhere (`curves/`, `surfaces/`, `structures/`). Rename them for consistency; no behavior change.

- move `include/igesio/numerics/mesh/**` → `numerics/meshes/**` (triangle_mesh.h and algorithms.h + algorithms/{conversion,edges,inspection, normals,mesh_line_intersection}.h)
- move `include/igesio/entities/mesh_entity.h` → `entities/meshes/mesh_entity.h` and `src/entities/mesh_entity.cpp` → `entities/meshes/mesh_entity.cpp`, with the test moved to `tests/entities/meshes/test_mesh_entity.cpp`
- rename the include guards to match the new paths
- update all include sites (the numerics.h / entities.h aggregates, graphics, the STL/OBJ extensions, and tests) and the entities CMake source lists
