+++ include/igesio/graphics/core/mesh_staging.h
@@ -16,7 +16,7 @@
 #include <cstddef>
 #include <vector>
 
-#include "igesio/numerics/mesh/triangle_mesh.h"
+#include "igesio/numerics/meshes/triangle_mesh.h"
