+++ tests/extensions/test_stl_io.cpp
@@ -21,8 +21,8 @@
 #include <string>
 
 #include "igesio/common/errors.h"
-#include "igesio/numerics/mesh/triangle_mesh.h"
-#include "igesio/numerics/mesh/algorithms.h"
+#include "igesio/numerics/meshes/triangle_mesh.h"
+#include "igesio/numerics/meshes/algorithms.h"
 #include "igesio/extensions/stl/stl_io.h"
 
 namespace {
