+++ tests/graphics/test_mesh_staging.cpp
@@ -19,7 +19,7 @@
 
 #include <vector>
 
-#include "igesio/numerics/mesh/triangle_mesh.h"
+#include "igesio/numerics/meshes/triangle_mesh.h"
 #include "igesio/graphics/core/mesh_staging.h"
 
 namespace {
