+++ tests/numerics/test_mesh_line_intersection.cpp
@@ -17,8 +17,8 @@
 
 #include <vector>
 
-#include "igesio/numerics/mesh/triangle_mesh.h"
-#include "igesio/numerics/mesh/algorithms.h"
+#include "igesio/numerics/meshes/triangle_mesh.h"
+#include "igesio/numerics/meshes/algorithms.h"
 
 namespace {
